fix: remove CustomTemplatesDir related codes
